Refactor profile edit and badges into dedicated modals, add level indicator to ProfileCard (#251, #252, #253)
