+++ app/_layout.tsx
@@ -26,7 +26,7 @@ function RootNavigator() {
     if (isLoading) return;
 
     const onboardingComplete = preferences.onboardingComplete === true;
-    const inApp = segments[0] === '(tabs)' || segments[0] === 'add-food-modal' || segments[0] === 'create-meal-modal' || segments[0] === 'app-settings-modal' || segments[0] === 'appearance-modal' || segments[0] === 'nutrition-goals-modal' || segments[0] === 'weekly-recap-modal' || segments[0] === 'food-library-modal';
+    const inApp = segments[0] === '(tabs)' || segments[0] === 'add-food-modal' || segments[0] === 'create-meal-modal' || segments[0] === 'app-settings-modal' || segments[0] === 'appearance-modal' || segments[0] === 'nutrition-goals-modal' || segments[0] === 'weekly-recap-modal' || segments[0] === 'food-library-modal' || segments[0] === 'profile-modal' || segments[0] === 'stats-achievements-modal';
 
     if (onboardingComplete && !inApp) {
       const tab = preferences.defaultTab ?? 'nutrition';
@@ -125,6 +125,20 @@ function RootNavigator() {
       />
       <Stack.Screen
         name="food-library-modal"
+        options={{
+          presentation: 'modal',
+          headerShown: false,
+        }}
+      />
+      <Stack.Screen
+        name="profile-modal"
+        options={{
+          presentation: 'modal',
+          headerShown: false,
+        }}
+      />
+      <Stack.Screen
+        name="stats-achievements-modal"
         options={{
           presentation: 'modal',
           headerShown: false,
